tests for map io

+++ DDNetLib/include/ddnet/util/utility.hpp
@@ -63,11 +63,11 @@ inline constexpr _UnderlyingType toUnderlying(_EnumType value) noexcept {
 }
 
 inline constexpr qint32 floatToFixed(qreal value) noexcept { 
-    return static_cast<int>(value * static_cast<qreal>(1 << constants::_fractional_bits));
+    return static_cast<qint32>(value * static_cast<qreal>(1 << constants::_fractional_bits));
 }
 
-inline constexpr float fixedToFloat(qint32 value) noexcept { 
-    return value * (1.f / (1 << constants::_fractional_bits));
+inline constexpr qreal fixedToFloat(qint32 value) noexcept {
+    return value * (1.0 / (1 << constants::_fractional_bits));
 }
 
 } // ddnet::util::
